Refactoring: Modularize CrateCompiler
By refactoring the SQLAlchemy 1.0-1.3 vs. 1.4 specialities of the
`CrateCompiler` (`visit_update` and `_get_crud_params`) into dedicated
modules, the new layout is more suitable for bringing in support for
SQLAlchemy 2.0.
